Use 32-bit parameter buffers

+++ src/buffer/audio_buffer.rs
@@ -141,7 +141,7 @@ pub trait AudioBuffer {
         }
     }
 
-    fn apply_gain(&mut self, gain: &[f64]) {
+    fn apply_gain(&mut self, gain: &[f32]) {
         debug_assert_eq!(gain.len(), self.frame_count());
 
         if gain.iter().all(|gain| gain.abs() < 1e-9) {
@@ -157,7 +157,7 @@ pub trait AudioBuffer {
             let channel_data = self.get_channel_data_mut(SampleLocation::new(channel, 0));
 
             for (sample, gain) in channel_data.iter_mut().zip(gain.iter()) {
-                *sample *= *gain as f32;
+                *sample *= *gain;
             }
         }
     }
